Worked on volumetric clouds
- Accessible now via debug menu

+++ shaders/lib/Settings.glsl
@@ -20,7 +20,7 @@ const int colortex1Format = R11F_G11F_B10F;
 const int colortex4Format = RG32F;
 
 ** composite0 Buffers **
-const int colortex5Format = RGBA8;
+const int colortex5Format = RGBA16;
 const int colortex6Format = RG8;
